+++ src/pages/create/components/LedgerAddressIndexBalanceTable.tsx
@@ -49,9 +49,9 @@ const LedgerAddressIndexBalanceTable = (props: {
   const [t] = useTranslation();
 
   const network = props.form?.getFieldValue('network');
-  const isTestnet = network === DefaultWalletConfigs.TestNetCroeseid4Config.name;
+  const isTestnet = network === DefaultWalletConfigs.TestNetCroeseid5Config.name;
   const config = isTestnet
-    ? DefaultWalletConfigs.TestNetCroeseid4Config
+    ? DefaultWalletConfigs.TestNetCroeseid5Config
     : DefaultWalletConfigs.MainNetConfig;
   const cronosTendermintAsset = {
     ...CRONOS_TENDERMINT_ASSET(config),
@@ -169,11 +169,13 @@ const LedgerAddressIndexBalanceTable = (props: {
     setLoading(true);
     switch (`${assetType}-${chainName}`) {
       case `${UserAssetType.TENDERMINT}-${SupportedChainName.CRYPTO_ORG}`: {
-        const nodeUrl = isTestnet
-          ? DefaultWalletConfigs.TestNetCroeseid4Config.nodeUrl
-          : DefaultWalletConfigs.MainNetConfig.nodeUrl;
+        const config = isTestnet
+          ? DefaultWalletConfigs.TestNetCroeseid5Config
+          : DefaultWalletConfigs.MainNetConfig;
         const nodeRpc = await NodeRpcService.init({
-          baseUrl: nodeUrl,
+          baseUrl: config.nodeUrl,
+          clientUrl: config.tendermintNetwork?.node?.clientUrl,
+          proxyUrl: config.tendermintNetwork?.node?.proxyUrl,
         });
         await Promise.all(
           ledgerAccountList.map(async account => {