+++ src/pages/create/create.tsx
@@ -29,7 +29,6 @@ import {
   DefaultWalletConfigs,
   LedgerWalletMaximum,
   NetworkName,
-  NodePorts,
   SupportedChainName,
 } from '../../config/StaticConfig';
 import logo from '../../assets/logo-products-chain.svg';
@@ -138,7 +137,7 @@ const FormCustomConfig: React.FC<FormCustomConfigProps> = (props) => {
       setCheckingNodeConnection(true);
       const { nodeUrl } = values;
       const isNodeLive = await walletService.checkNodeIsLive(
-        `${nodeUrl}${NodePorts.Tendermint}`,
+        `${nodeUrl}`,
       );
       setCheckingNodeConnection(false);
 
@@ -800,7 +799,7 @@ const FormCreate: React.FC<FormCreateProps> = (props) => {
         case `${UserAssetType.TENDERMINT}-${SupportedChainName.CRYPTO_ORG}`:
           {
             const addressPrefix =
-              network === DefaultWalletConfigs.TestNetCroeseid4Config.name
+              network === DefaultWalletConfigs.TestNetCroeseid5Config.name
                 ? 'tcro'
                 : 'cro';
             const tendermintAddressList = await device.getAddressList(