+++ src/pages/governance/governance.tsx
@@ -15,6 +15,7 @@ import {
   Input,
   InputNumber,
   Progress,
+  Spin,
 } from 'antd';
 
 import Big from 'big.js';
@@ -25,12 +26,13 @@ import {
   InfoCircleOutlined,
   ArrowLeftOutlined,
   FieldTimeOutlined,
+  LoadingOutlined,
 } from '@ant-design/icons';
-import { useRecoilState, useRecoilValue } from 'recoil';
+import { useRecoilState } from 'recoil';
 import { useTranslation } from 'react-i18next';
 import { ledgerIsExpertModeState, sessionState, walletAssetState } from '../../recoil/atom';
 
-import { getUIVoteAmount, getUIDynamicAmount, getBaseScaledAmount } from '../../utils/NumberUtils';
+import { getUIVoteAmount, getUIDynamicAmount, getBaseScaledAmount, getUINormalScaleAmount } from '../../utils/NumberUtils';
 
 import {
   ProposalModel,
@@ -55,6 +57,8 @@ import { middleEllipsis } from '../../utils/utils';
 
 import { ProposalView } from './components/ProposalView';
 import { VotingHistory } from './components/VotingHistory';
+import { useCronosTendermintAsset } from '../../hooks/useAsset';
+import BigNumber from 'bignumber.js';
 
 const { Header, Content, Footer } = Layout;
 const { TabPane } = Tabs;
@@ -81,6 +85,8 @@ const GovernancePage = () => {
   const [decryptedPhrase, setDecryptedPhrase] = useState('');
   const [errorMessages, setErrorMessages] = useState([]);
   const [proposal, setProposal] = useState<ProposalModel>();
+
+  const cronosTendermintAsset = useCronosTendermintAsset(); 
 
   const numWithCommas = (x: string) => {
     if (x) {
@@ -112,11 +118,12 @@ const GovernancePage = () => {
   const [userAsset, setUserAsset] = useRecoilState(walletAssetState);
   const [proposalList, setProposalList] = useState<ProposalModel[]>();
   const [isConfirmationModalVisible, setIsVisibleConfirmationModal] = useState(false);
-  const currentSession = useRecoilValue(sessionState);
+  const [currentSession, setCurrentSession] = useRecoilState(sessionState);
   const didMountRef = useRef(false);
   const [isLoadingTally, setIsLoadingTally] = useState(false);
   const minDeposit = '1000';
-  const maxDeposit = '10001';
+  const [passingDeposit, setPassingDeposit] = useState('10001');
+  const [isLoadingDeposit, setIsLoadingDeposit] = useState(true);
 
   const [createProposalHash, setCreateProposalHash] = useState('');
   const [initialDepositProposal, setInitialDeposit] = useState('0');
@@ -135,9 +142,9 @@ const GovernancePage = () => {
   );
 
   const customMaxValidator = TransactionUtils.maxValidator(
-    maxDeposit,
+    passingDeposit,
     t('governance.modal2.form.input.proposalDeposit.max.error', {
-      maxDeposit: numWithCommas(maxDeposit)
+      maxDeposit: numWithCommas(passingDeposit)
         .concat(' ')
         .concat(userAsset?.symbol),
     }),
@@ -200,6 +207,10 @@ const GovernancePage = () => {
   };
 
   const checkProposalType = (proposal: any) => {
+    if (!proposal || !proposal?.content) {
+      return '';
+    }
+    
     const proposal_initial_deposit = proposal?.content['@type']?.toLowerCase();
     let status = '';
     switch (true) {
@@ -464,7 +475,7 @@ const GovernancePage = () => {
     try {
       if (
         Big(form?.getFieldValue('initialDeposit')).cmp(Big(minDeposit)) !== -1 &&
-        Big(form?.getFieldValue('initialDeposit')).cmp(Big(maxDeposit)) !== 1
+        Big(form?.getFieldValue('initialDeposit')).cmp(Big(passingDeposit)) !== 1
       ) {
         setConfirmLoading(true);
         const proposalType = form.getFieldValue('proposalType');
@@ -572,7 +583,7 @@ const GovernancePage = () => {
       .toString();
     const totalDeposit = Big(getUIDynamicAmount(depositCalc, userAsset)).toString();
     const finalPercentage = Big(totalDeposit)
-      .div(Big(maxDeposit.replace(',', '')))
+      .div(Big(passingDeposit.replace(',', '')))
       .times(100)
       .toFixed(2);
     return Big(finalPercentage).toNumber();
@@ -601,12 +612,29 @@ const GovernancePage = () => {
     }, 300);
   };
 
+  const fetchProposalDeposit = async () => {
+    setIsLoadingDeposit(true);
+    const deposit = await walletService.fetchProposalMinDeposit() ?? '2000000000000';
+    const adjustedDeposit = new BigNumber(deposit).plus('100000000').toString(); // Plus 1 CRO
+    setPassingDeposit(getUINormalScaleAmount(adjustedDeposit, userAsset.decimals, 2));
+    setIsLoadingDeposit(false);
+  };
+
   useEffect(() => {
     fetchProposalList();
 
     if (!didMountRef.current) {
       const usersBalance = getUIDynamicAmount(userAsset?.balance, userAsset);
       const userDeposit = Big(usersBalance).cmp(Big(minDeposit)) === 1 ? minDeposit : usersBalance;
+      setCurrentSession({
+        ...currentSession,
+        activeAsset: cronosTendermintAsset,
+      });
+      walletService.setCurrentSession({
+        ...currentSession,
+        activeAsset: cronosTendermintAsset,
+      });
+      fetchProposalDeposit();
       didMountRef.current = true;
       analyticsService.logPage('Governance');
       form.setFieldsValue({ initialDeposit: userDeposit });
@@ -614,7 +642,7 @@ const GovernancePage = () => {
     }
 
     // eslint-disable-next-line
-  }, [currentSession, form, userAsset, proposal, proposalList, setModalType, modalType]);
+  }, [currentSession, form, userAsset, proposal, proposalList, setModalType, modalType, isLoadingTally, isLoadingDeposit]);
 
   return (
     <Layout className="site-layout">
@@ -622,7 +650,7 @@ const GovernancePage = () => {
         <>
           <Header className="site-layout-background proposal-details-header">
             <div className="top">{t('governance.proposalDetails')}</div>
-            {proposal?.content.title}
+            {proposal?.content?.title}
           </Header>
 
           <a className="proposal-back-btn">
@@ -651,7 +679,7 @@ const GovernancePage = () => {
         <>
           <Header className="site-layout-background proposal-details-header">
             <div className="top">{t('governance.proposalDetails')}</div>
-            {proposal?.content.title}
+            {proposal?.content?.title}
           </Header>
 
           <a className="proposal-back-btn">
@@ -982,6 +1010,8 @@ const GovernancePage = () => {
               // setProposal
               proposalList,
               setProposalList,
+              passingDeposit,
+              isLoadingDeposit
             }}
           />
         ) : (
@@ -999,6 +1029,12 @@ const GovernancePage = () => {
                   <div className="site-layout-background governance-content">
                     <div className="container">
                       <List
+                        loading={{
+                          indicator: (
+                            <Spin indicator={<LoadingOutlined style={{ fontSize: 36 }} spin />} />
+                          ),
+                          spinning: isLoadingDeposit || isLoadingTally,
+                        }}
                         dataSource={proposalList}
                         renderItem={(item: ProposalModel) => (
                           <List.Item
@@ -1034,7 +1070,7 @@ const GovernancePage = () => {
                               title={
                                 <>
                                   {processStatusTag(item.status)} #{item.proposal_id}{' '}
-                                  <a>{item.content.title}</a>
+                                  <a>{item.content?.title}</a>
                                 </>
                               }
                               description={
@@ -1135,7 +1171,7 @@ const GovernancePage = () => {
                               title={
                                 <>
                                   {processStatusTag(item.status)} #{item.proposal_id}{' '}
-                                  <a>{item.content.title}</a>
+                                  <a>{item.content?.title}</a>
                                 </>
                               }
                               description={
@@ -1191,7 +1227,7 @@ const GovernancePage = () => {
                               title={
                                 <>
                                   {processStatusTag(item.status)} #{item.proposal_id}{' '}
-                                  <a>{item.content.title}</a>
+                                  <a>{item.content?.title}</a>
                                 </>
                               }
                               description={
@@ -1247,7 +1283,7 @@ const GovernancePage = () => {
                               title={
                                 <>
                                   {processStatusTag(item.status)} #{item.proposal_id}{' '}
-                                  <a>{item.content.title}</a>
+                                  <a>{item.content?.title}</a>
                                 </>
                               }
                               description={
@@ -1303,7 +1339,7 @@ const GovernancePage = () => {
                               title={
                                 <>
                                   {processStatusTag(item.status)} #{item.proposal_id}{' '}
-                                  <a>{item.content.title}</a>
+                                  <a>{item.content?.title}</a>
                                 </>
                               }
                               description={
@@ -1359,7 +1395,7 @@ const GovernancePage = () => {
                               title={
                                 <>
                                   {processStatusTag(item.status)} #{item.proposal_id}{' '}
-                                  <a>{item.content.title}</a>
+                                  <a>{item.content?.title}</a>
                                 </>
                               }
                               description={
@@ -1430,7 +1466,7 @@ const GovernancePage = () => {
           </div>
           <div className="item">
             <div className="label">{t('governance.modal1.label2')}</div>
-            <div className="address">{`#${proposal?.proposal_id} ${proposal?.content.title}`}</div>
+            <div className="address">{`#${proposal?.proposal_id} ${proposal?.content?.title}`}</div>
           </div>
           <div className="item">
             <div className="label">{t('governance.modal1.label3')}</div>