+++ src/pages/settings/components/FormSettings.tsx
@@ -11,8 +11,8 @@ import { walletService } from '../../../service/WalletService';
 import { SettingsDataUpdate } from '../../../models/Wallet';
 import ModalPopup from '../../../components/ModalPopup/ModalPopup';
 
-import { FIXED_DEFAULT_FEE, FIXED_DEFAULT_GAS_LIMIT } from '../../../config/StaticConfig';
-import { UserAsset, UserAssetConfig } from '../../../models/UserAsset';
+import { FIXED_DEFAULT_FEE, FIXED_DEFAULT_GAS_LIMIT, SupportedChainName, TestNetCroeseid4Config, TestNetCroeseid5Config } from '../../../config/StaticConfig';
+import { UserAsset, UserAssetConfig, UserAssetType } from '../../../models/UserAsset';
 import AddressBook from '../tabs/AddressBook/AddressBook';
 import { getCronosTendermintAsset } from '../../../utils/utils';
 import RevokePermission from '../tabs/RevokePermission/RevokePermission';
@@ -73,9 +73,12 @@ export const FormSettings = () => {
       nodeUrl: prevSettings.nodeUrl,
       chainId: prevSettings.chainId,
       indexingUrl: prevSettings.indexingUrl,
+      clientUrl: prevSettings.tendermintNetwork?.node?.clientUrl,
+      proxyUrl: prevSettings.tendermintNetwork?.node?.proxyUrl,
       networkFee,
       gasLimit,
     });
+
     if (!currentAssetIdentifier && croAsset) {
       setCurrentAssetIdentifier(croAsset?.identifier);
 
@@ -99,6 +102,8 @@ export const FormSettings = () => {
     if (
       prevSettings.nodeUrl === values.nodeUrl &&
       prevSettings.indexingUrl === values.indexingUrl &&
+      prevSettings.tendermintNetwork?.node?.clientUrl === values.clientUrl &&
+      prevSettings.tendermintNetwork?.node?.proxyUrl === values.proxyUrl &&
       prevSettings.chainId === values.chainId &&
       defaultGasLimit === values.gasLimit &&
       defaultNetworkFee === values.networkFee
@@ -113,13 +118,26 @@ export const FormSettings = () => {
       chainId: values.chainId,
       nodeUrl: values.nodeUrl,
       indexingUrl: values.indexingUrl,
+      clientUrl: values.clientUrl,
+      proxyUrl: values.proxyUrl,
       networkFee: String(values.networkFee),
       gasLimit: String(values.gasLimit),
     };
 
     // This wallet level settings update should only imply the primary asset.
-    if (!session.activeAsset?.isSecondaryAsset) {
+    if (
+      session.activeAsset?.assetType === UserAssetType.TENDERMINT &&
+      session.activeAsset?.name === SupportedChainName.CRYPTO_ORG
+    ) {
       await walletService.updateWalletNodeConfig(settingsDataUpdate);
+    }
+
+    if (settingsDataUpdate.chainId === TestNetCroeseid4Config.network.chainId) {
+      settingsDataUpdate.explorer = TestNetCroeseid4Config.explorer;
+    }
+
+    if (settingsDataUpdate.chainId === TestNetCroeseid5Config.network.chainId) {
+      settingsDataUpdate.explorer = TestNetCroeseid5Config.explorer;
     }
 
     const updatedWallet = await walletService.findWalletByIdentifier(session.wallet.identifier);
@@ -134,6 +152,22 @@ export const FormSettings = () => {
         fee: { gasLimit: settingsDataUpdate.gasLimit!, networkFee: settingsDataUpdate.networkFee! },
         indexingUrl: settingsDataUpdate.indexingUrl!,
         nodeUrl: settingsDataUpdate.nodeUrl!,
+        explorer: settingsDataUpdate.explorer,
+        ...(
+          settingsDataUpdate.chainId === TestNetCroeseid4Config.network.chainId || 
+          settingsDataUpdate.chainId === TestNetCroeseid5Config.network.chainId
+        ) && {
+          explorer: settingsDataUpdate.explorer
+        },
+        ...(settingsDataUpdate.clientUrl || settingsDataUpdate.proxyUrl) && {
+          tendermintNetwork: {
+            ...previousAssetConfig?.tendermintNetwork!,
+            node: {
+              clientUrl: settingsDataUpdate.clientUrl ?? '',
+              proxyUrl: settingsDataUpdate.proxyUrl ?? '',
+            }
+          }
+        }
       },
     };
 
@@ -154,6 +188,7 @@ export const FormSettings = () => {
 
     const allAssets = await walletService.retrieveCurrentWalletAssets(newSession);
     setWalletAllAssets(allAssets);
+    walletService.syncAll();
 
     setIsButtonLoading(false);
     message.success(
@@ -187,6 +222,8 @@ export const FormSettings = () => {
       indexingUrl: defaultAssetConfig?.config.indexingUrl,
       networkFee: defaultAssetConfig?.config.fee.networkFee,
       gasLimit: defaultAssetConfig?.config.fee.gasLimit,
+      clientUrl: (session.activeAsset?.assetType === UserAssetType.TENDERMINT) ? defaultAssetConfig?.config.tendermintNetwork?.node?.clientUrl : '',
+      proxyUrl: (session.activeAsset?.assetType === UserAssetType.TENDERMINT) ? defaultAssetConfig?.config.tendermintNetwork?.node?.proxyUrl : '',
     });
   };
 
@@ -237,6 +274,7 @@ export const FormSettings = () => {
                       <GeneralSettingsForm
                         currentAssetIdentifier={currentAssetIdentifier}
                         setCurrentAssetIdentifier={setCurrentAssetIdentifier}
+                        form={form}
                       />
                       <Form.Item {...tailLayout} className="button">
                         <Button type="primary" htmlType="submit" loading={isButtonLoading}>